+++ web/apps/photos/src/components/Collections/CollectionShare/index.tsx
@@ -1,11 +1,16 @@
 import { SidebarDrawer } from "@/base/components/mui/SidebarDrawer";
 import { Titlebar } from "@/base/components/Titlebar";
-import type { Collection } from "@/media/collection";
+import { useModalVisibility } from "@/base/components/utils/modal";
+import type { Collection, PublicURL } from "@/media/collection";
+import { PublicLinkCreated } from "@/new/photos/components/share/PublicLinkCreated";
 import type { CollectionSummary } from "@/new/photos/services/collection/ui";
 import { DialogProps, Stack } from "@mui/material";
 import { t } from "i18next";
+import { useEffect, useState } from "react";
+import { appendCollectionKeyToShareURL } from "utils/collection";
 import EmailShare from "./emailShare";
-import PublicShare from "./publicShare";
+import EnablePublicShareOptions from "./publicShare/EnablePublicShareOptions";
+import { ManagePublicShare } from "./publicShare/managePublicShare";
 import SharingDetails from "./sharingDetails";
 
 interface CollectionShareProps {
@@ -79,5 +84,69 @@ export const CollectionShare: React.FC<CollectionShareProps> = ({
                 </Stack>
             </Stack>
         </SidebarDrawer>
+    );
+};
+
+interface PublicShareProps {
+    collection: Collection;
+    onRootClose: () => void;
+}
+
+const PublicShare: React.FC<PublicShareProps> = ({
+    collection,
+    onRootClose,
+}) => {
+    const [publicShareUrl, setPublicShareUrl] = useState<string>(null);
+    const [publicShareProp, setPublicShareProp] = useState<PublicURL>(null);
+    const {
+        show: showPublicLinkCreated,
+        props: publicLinkCreatedVisibilityProps,
+    } = useModalVisibility();
+
+    useEffect(() => {
+        if (collection.publicURLs?.length) {
+            setPublicShareProp(collection.publicURLs[0]);
+        }
+    }, [collection]);
+
+    useEffect(() => {
+        if (publicShareProp) {
+            const url = appendCollectionKeyToShareURL(
+                publicShareProp.url,
+                collection.key,
+            );
+            setPublicShareUrl(url);
+        } else {
+            setPublicShareUrl(null);
+        }
+    }, [publicShareProp]);
+
+    const copyToClipboardHelper = () => {
+        navigator.clipboard.writeText(publicShareUrl);
+    };
+
+    return (
+        <>
+            {publicShareProp ? (
+                <ManagePublicShare
+                    publicShareProp={publicShareProp}
+                    setPublicShareProp={setPublicShareProp}
+                    collection={collection}
+                    publicShareUrl={publicShareUrl}
+                    onRootClose={onRootClose}
+                    copyToClipboardHelper={copyToClipboardHelper}
+                />
+            ) : (
+                <EnablePublicShareOptions
+                    setPublicShareProp={setPublicShareProp}
+                    collection={collection}
+                    onLinkCreated={showPublicLinkCreated}
+                />
+            )}
+            <PublicLinkCreated
+                {...publicLinkCreatedVisibilityProps}
+                onCopyLink={copyToClipboardHelper}
+            />
+        </>
     );
 };
